+++ GPU/GPUTracking/Global/GPUChainTracking.h
@@ -97,7 +97,7 @@ class GPUChainTracking : public GPUChain, GPUReconstructionHelpers::helperDelega
     InOutMemory(InOutMemory&&);
     InOutMemory& operator=(InOutMemory&&);
 
-    std::unique_ptr<unsigned long long int[]> tpcZSpages;
+    std::unique_ptr<unsigned long[]> tpcZSpages;
     std::unique_ptr<char[]> tpcZSpagesChar;        // Same as above, but as char (needed for reading dumps, but deprecated, since alignment can be wrong) // TODO: Fix alignment
     std::unique_ptr<char[]> tpcCompressedClusters; // TODO: Fix alignment
     std::unique_ptr<GPUTrackingInOutZS> tpcZSmeta;
